:sparkles: generic combination generator
