+++ src/util/util.hpp
@@ -14,6 +14,8 @@
 #include <optional>
 #include <span>
 #include <string>
+#include <tl/generator.hpp>
+#include <tl/to.hpp>
 #include <vector>
 
 class tqdm;
@@ -125,6 +127,56 @@ overloaded(Ts...) -> overloaded<Ts...>;
 
 auto contains(std::ranges::range auto r, auto const& value) -> bool {
     return std::ranges::find(r, value) != r.end();
+}
+
+template <std::bidirectional_iterator InputIt>
+bool next_combination(const InputIt first, InputIt k, const InputIt last) {
+    /* Credits: Mark Nelson http://marknelson.us */
+    if ((first == last) || (first == k) || (last == k))
+        return false;
+    InputIt i1 = first;
+    InputIt i2 = last;
+    ++i1;
+    if (last == i1)
+        return false;
+    i1 = last;
+    --i1;
+    i1 = k;
+    --i2;
+    while (first != i1) {
+        if (*--i1 < *i2) {
+            InputIt j = k;
+            while (!(*i1 < *j)) ++j;
+            std::iter_swap(i1, j);
+            ++i1;
+            ++j;
+            i2 = k;
+            std::rotate(i1, j, last);
+            while (last != j) {
+                ++j;
+                ++i2;
+            }
+            std::rotate(k, i2, last);
+            return true;
+        }
+    }
+    std::rotate(first, k, last);
+    return false;
+}
+
+template <std::ranges::bidirectional_range R>
+auto next_combination(R& r, size_t const comb_size) {
+    return next_combination(r.begin(), dvlab::iterator::next(r.begin(), comb_size), r.end());
+}
+
+// iterate over all combinations in the range
+template <typename T>
+tl::generator<std::vector<T>> combinations(std::vector<T> elements, size_t const comb_size) {
+    std::ranges::sort(elements);
+    do {  // NOLINT(cppcoreguidelines-avoid-do-while)
+        auto comb = elements | std::views::take(comb_size) | tl::to<std::vector<T>>();
+        co_yield comb;
+    } while (next_combination(elements, comb_size));
 }
 
 }  // namespace dvlab
