+++ tests/simulation/tof3/ref/assign2not.log
@@ -25,25 +25,25 @@ ID:   17 (Z, 0)       (Qubit, Col): (-1, 5)        #Neighbors:   3    (11, H) (1
 ID:   18 (Z, -π/4)    (Qubit, Col): (-2, 5)        #Neighbors:   1    (17, H)
 ID:   19 (Z, 0)       (Qubit, Col): (-1, 5)        #Neighbors:   4    (10, H) (11, H) (12, H) (20, H)
 ID:   20 (Z, π/4)     (Qubit, Col): (-2, 5)        #Neighbors:   1    (19, H)
-ID:   21 (Z, π/4)     (Qubit, Col): (2, 10)        #Neighbors:   5    (4, -) (5, -) (24, H) (26, H) (30, H)
+ID:   21 (Z, π/4)     (Qubit, Col): (2, 10)        #Neighbors:   5    (4, -) (5, -) (26, H) (28, H) (30, H)
 ID:   22 (Z, π/4)     (Qubit, Col): (4, 10)        #Neighbors:   5    (12, H) (24, H) (28, H) (30, H) (34, H)
-ID:   23 (Z, π/4)     (Qubit, Col): (3, 10)        #Neighbors:   5    (6, H) (7, H) (26, H) (28, H) (30, H)
-ID:   24 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (21, H) (22, H) (25, H)
+ID:   23 (Z, π/4)     (Qubit, Col): (3, 10)        #Neighbors:   5    (6, H) (7, H) (24, H) (26, H) (30, H)
+ID:   24 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (22, H) (23, H) (25, H)
 ID:   25 (Z, -π/4)    (Qubit, Col): (-2, 10)       #Neighbors:   1    (24, H)
 ID:   26 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (21, H) (23, H) (27, H)
 ID:   27 (Z, -π/4)    (Qubit, Col): (-2, 10)       #Neighbors:   1    (26, H)
-ID:   28 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (22, H) (23, H) (29, H)
+ID:   28 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (21, H) (22, H) (29, H)
 ID:   29 (Z, -π/4)    (Qubit, Col): (-2, 10)       #Neighbors:   1    (28, H)
 ID:   30 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   4    (21, H) (22, H) (23, H) (31, H)
 ID:   31 (Z, π/4)     (Qubit, Col): (-2, 10)       #Neighbors:   1    (30, H)
-ID:   32 (Z, π/4)     (Qubit, Col): (0, 15)        #Neighbors:   5    (1, -) (10, -) (35, H) (37, H) (41, H)
+ID:   32 (Z, π/4)     (Qubit, Col): (0, 15)        #Neighbors:   5    (1, -) (10, -) (37, H) (39, H) (41, H)
 ID:   33 (Z, π/4)     (Qubit, Col): (1, 15)        #Neighbors:   5    (3, -) (11, -) (35, H) (39, H) (41, H)
-ID:   34 (Z, π/4)     (Qubit, Col): (4, 15)        #Neighbors:   5    (9, H) (22, H) (37, H) (39, H) (41, H)
-ID:   35 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (32, H) (33, H) (36, H)
+ID:   34 (Z, π/4)     (Qubit, Col): (4, 15)        #Neighbors:   5    (9, H) (22, H) (35, H) (37, H) (41, H)
+ID:   35 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (33, H) (34, H) (36, H)
 ID:   36 (Z, -π/4)    (Qubit, Col): (-2, 15)       #Neighbors:   1    (35, H)
 ID:   37 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (32, H) (34, H) (38, H)
 ID:   38 (Z, -π/4)    (Qubit, Col): (-2, 15)       #Neighbors:   1    (37, H)
-ID:   39 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (33, H) (34, H) (40, H)
+ID:   39 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (32, H) (33, H) (40, H)
 ID:   40 (Z, -π/4)    (Qubit, Col): (-2, 15)       #Neighbors:   1    (39, H)
 ID:   41 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   4    (32, H) (33, H) (34, H) (42, H)
 ID:   42 (Z, π/4)     (Qubit, Col): (-2, 15)       #Neighbors:   1    (41, H)
@@ -81,25 +81,25 @@ ID:   17 (Z, 0)       (Qubit, Col): (-1, 5)        #Neighbors:   3    (11, H) (1
 ID:   18 (Z, -π/4)    (Qubit, Col): (-2, 5)        #Neighbors:   1    (17, H)
 ID:   19 (Z, 0)       (Qubit, Col): (-1, 5)        #Neighbors:   4    (10, H) (11, H) (12, H) (20, H)
 ID:   20 (Z, π/4)     (Qubit, Col): (-2, 5)        #Neighbors:   1    (19, H)
-ID:   21 (Z, π/4)     (Qubit, Col): (2, 10)        #Neighbors:   5    (24, H) (26, H) (30, H) (47, -) (48, -)
+ID:   21 (Z, π/4)     (Qubit, Col): (2, 10)        #Neighbors:   5    (26, H) (28, H) (30, H) (47, -) (48, -)
 ID:   22 (Z, π/4)     (Qubit, Col): (4, 10)        #Neighbors:   5    (12, H) (24, H) (28, H) (30, H) (34, H)
-ID:   23 (Z, π/4)     (Qubit, Col): (3, 10)        #Neighbors:   5    (6, H) (7, H) (26, H) (28, H) (30, H)
-ID:   24 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (21, H) (22, H) (25, H)
+ID:   23 (Z, π/4)     (Qubit, Col): (3, 10)        #Neighbors:   5    (6, H) (7, H) (24, H) (26, H) (30, H)
+ID:   24 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (22, H) (23, H) (25, H)
 ID:   25 (Z, -π/4)    (Qubit, Col): (-2, 10)       #Neighbors:   1    (24, H)
 ID:   26 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (21, H) (23, H) (27, H)
 ID:   27 (Z, -π/4)    (Qubit, Col): (-2, 10)       #Neighbors:   1    (26, H)
-ID:   28 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (22, H) (23, H) (29, H)
+ID:   28 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   3    (21, H) (22, H) (29, H)
 ID:   29 (Z, -π/4)    (Qubit, Col): (-2, 10)       #Neighbors:   1    (28, H)
 ID:   30 (Z, 0)       (Qubit, Col): (-1, 10)       #Neighbors:   4    (21, H) (22, H) (23, H) (31, H)
 ID:   31 (Z, π/4)     (Qubit, Col): (-2, 10)       #Neighbors:   1    (30, H)
-ID:   32 (Z, π/4)     (Qubit, Col): (0, 15)        #Neighbors:   5    (10, -) (35, H) (37, H) (41, H) (44, -)
+ID:   32 (Z, π/4)     (Qubit, Col): (0, 15)        #Neighbors:   5    (10, -) (37, H) (39, H) (41, H) (44, -)
 ID:   33 (Z, π/4)     (Qubit, Col): (1, 15)        #Neighbors:   5    (11, -) (35, H) (39, H) (41, H) (46, -)
-ID:   34 (Z, π/4)     (Qubit, Col): (4, 15)        #Neighbors:   5    (22, H) (37, H) (39, H) (41, H) (50, H)
-ID:   35 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (32, H) (33, H) (36, H)
+ID:   34 (Z, π/4)     (Qubit, Col): (4, 15)        #Neighbors:   5    (22, H) (35, H) (37, H) (41, H) (50, H)
+ID:   35 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (33, H) (34, H) (36, H)
 ID:   36 (Z, -π/4)    (Qubit, Col): (-2, 15)       #Neighbors:   1    (35, H)
 ID:   37 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (32, H) (34, H) (38, H)
 ID:   38 (Z, -π/4)    (Qubit, Col): (-2, 15)       #Neighbors:   1    (37, H)
-ID:   39 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (33, H) (34, H) (40, H)
+ID:   39 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   3    (32, H) (33, H) (40, H)
 ID:   40 (Z, -π/4)    (Qubit, Col): (-2, 15)       #Neighbors:   1    (39, H)
 ID:   41 (Z, 0)       (Qubit, Col): (-1, 15)       #Neighbors:   4    (32, H) (33, H) (34, H) (42, H)
 ID:   42 (Z, π/4)     (Qubit, Col): (-2, 15)       #Neighbors:   1    (41, H)
